test(mcp-cli): split the 1640-line CLI test file for shard balance
mcp-cli.test.ts was 38% of the suite's wall time (one 24.7s file) because it
spawns the real CLI as a Node subprocess per test. Vitest shards by whole file,
so that one file landed entirely on a single shard and was the cause of the
75s/57s shard imbalance — adding more shards couldn't help because the file
couldn't be split.

Pure reorganization (no test logic, assertions, or titles changed); the 53 tests
are preserved and now distribute across shards/workers:

- test/unit/support/mcp-cli-harness.ts: shared helpers (run, runAsync, git,
  fixtures, startFixtureServer) extracted with exports. `server` is now private
  to the harness; new closeFixtureServer() replaces the in-test server teardown.
  Non-`.test.ts` name under support/ so the test glob does not collect it.
- test/unit/mcp-cli-basics.test.ts (16): snippets, version, completion, config.
- test/unit/mcp-cli-doctor.test.ts (14): doctor/status/compatibility/install.
- test/unit/mcp-cli-profiles.test.ts (6): profiles/login/logout + package status.
- test/unit/mcp-cli-packets.test.ts (17): decision packs + agent packets.

Verified: the 4 files total 53 passed (== the original), typecheck clean.

This redistributes the load; the larger win (in-process testing via an exported
runCli to drop the per-test subprocess spawn) is a follow-up that touches the
published binary and will land separately with subprocess smoke tests retained.
